docs: add readmes to new extensions
