+++ extensions/common/aas-lib/build.gradle.kts
@@ -5,11 +5,8 @@ plugins {
 dependencies {
     implementation(project(":extensions:common:constants"))
     implementation(project(":extensions:common:data-plane-aas-spi"))
-    implementation(project(":extensions:common:validator:validator-data-address-aas-data"))
+    runtimeOnly(project(":extensions:common:validator:validator-data-address-aas-data"))
     implementation(libs.edc.asset.spi)
-    implementation(libs.edc.data.plane.http.spi)
-    implementation(libs.aas4j.model)
-    implementation(libs.aas4j.dataformat.json)
     implementation(libs.fa3st.model) // ReferenceHelper
 
     testImplementation(libs.edc.junit)
