+++ extensions/common/constants/README.md
@@ -1,0 +1,15 @@
+# AAS Constants
+
+This module provides AAS constants and API paths.
+
+## Configuration
+
+No configuration variables.
+
+## Interfaces
+
+No outward-facing API.
+
+### Dependencies
+
+No dependencies.
