+++ extensions/common/data-plane-aas-spi/README.md
@@ -1,0 +1,20 @@
+# AAS Data-Plane SPI
+
+This module provides the AAS data address.
+
+## Configuration
+
+No configuration variables.
+
+## Interfaces
+
+No outward-facing API.
+
+### Dependencies
+
+| Name                                                | Description             |
+|:----------------------------------------------------|:------------------------|
+| common:constants (local)                            | AAS constants           |
+| org.eclipse.edc:data-plane-http-spi                 | HTTP BASE_URL field     |
+| org.eclipse.edc:core-spi                            | DataAddress class       |
+| org.eclipse.digitaltwin.aas4j:aas4j-dataformat-json | Json(De)Serializer      |
