+++ extensions/common/validator/validator-data-address-aas-data/README.md
@@ -1,0 +1,19 @@
+# AAS Data DataAddress Validator
+
+This extension provides the AAS Data DataAddress Validator for source and destination AAS-DAs.
+
+## Configuration
+
+No configuration variables.
+
+## Interfaces
+
+No outward-facing API.
+
+### Dependencies
+
+| Name                                | Description         |
+|:------------------------------------|:--------------------|
+| common:data-plane-aas-spi (local)   | AAS DataAddress     |
+| org.eclipse.edc:data-plane-http-spi | HTTP BASE_URL field |
+| org.eclipse.edc:validator-spi       | Register validator  |
